### 2021-10-26 #3855 @whedon
[PRE REVIEW]: PyMPDATA v1: Numba-accelerated Pythonic implementation of MPDATA with examples in Python, Julia and Matlab
**Submitting author:** @slayoo (<a href="http://orcid.org/0000-0003-2361-0082">Sylwester Arabas</a>)
**Repository:** <a href="https://github.com/atmos-cloud-sim-uj/PyMPDATA" target ="_blank">https://github.com/atmos-cloud-sim-uj/PyMPDATA</a>
**Version:** v1
