### 2021-10-27 5627b51
more pylint suggestions
