+++ PyMPDATA/impl/__init__.py
@@ -1,0 +1,2 @@
+""" PyMPDATA internals, if anything from within is needed to be referenced
+    from user code, please file a bug report on impl leaking into public API """
