### 2021-10-28 0700de6
and some more pylint suggestions
