+++ PyMPDATA/solver.py
@@ -21,8 +21,7 @@ class PostIterNull:
     def __init__(self):
         pass
 
-    # pylint: disable-next=unused-argument
-    def __call__(self, flux, g_factor, t, it):
+    def __call__(self, flux, g_factor, step, iteration):  # pylint: disable=unused-argument
         pass
 
 
