### 2021-10-28 54e6ecd
cleanups around solver & stepper
