### 2021-10-28 da1ad65
pylint suggestions + turning on fail-on-pylint-failure in GA workflow + add pylint badge
