+++ PyMPDATA/impl/formulae_antidiff.py
@@ -6,43 +6,43 @@ from PyMPDATA.impl.enumerations import MAX_DIM_NUM
 
 
 def make_antidiff(non_unit_g_factor, options, traversals, last_pass=False):
-    if options.n_iters <= 1:
-        @numba.njit(**options.jit_flags)
-        # pylint: disable=too-many-arguments
-        def apply(_1, _2, _3, _4, _5, _6, _7):
-            return
-    else:
-        idx = traversals.indexers[traversals.n_dims]
-        apply_vector = traversals.apply_vector()
+    """ returns njit-ted function for use with given traversals """
+    idx = traversals.indexers[traversals.n_dims]
+    apply_vector = traversals.apply_vector()
 
-        formulae_antidiff = tuple(
-            __make_antidiff(idx.atv[i], idx.ats[i],
-                            non_unit_g_factor=non_unit_g_factor,
-                            options=options,
-                            n_dims=traversals.n_dims,
-                            last_pass=last_pass)
-            if idx.ats[i] is not None else None
-            for i in range(MAX_DIM_NUM))
+    formulae_antidiff = tuple(
+        __make_antidiff(atv=idx.atv[i],
+                        ats=idx.ats[i],
+                        non_unit_g_factor=non_unit_g_factor,
+                        options=options,
+                        n_dims=traversals.n_dims,
+                        last_pass=last_pass)
+        if idx.ats[i] is not None else None
+        for i in range(MAX_DIM_NUM))
 
-        @numba.njit(**options.jit_flags)
-        # pylint: disable=too-many-arguments
-        def apply(GC_corr, psi, psi_bc, GC_unco, vec_bc, g_factor, g_factor_bc):
-            return apply_vector(*formulae_antidiff, *GC_corr, *psi, *psi_bc, *GC_unco, *vec_bc,
-                                *g_factor, *g_factor_bc)
+    @numba.njit(**options.jit_flags)
+    # pylint: disable=too-many-arguments
+    def apply(g_c_corr, psi, psi_bc, g_c_unco, vec_bc, g_factor, g_factor_bc):
+        return apply_vector(*formulae_antidiff,
+                            *g_c_corr,
+                            *psi, *psi_bc,
+                            *g_c_unco, *vec_bc,
+                            *g_factor, *g_factor_bc)
 
     return apply
 
 
-def __make_antidiff(atv, ats, non_unit_g_factor, options, n_dims, last_pass):
+# pylint: disable=too-many-locals
+def __make_antidiff(*, atv, ats, non_unit_g_factor, options, n_dims, last_pass):
     infinite_gauge = options.infinite_gauge
     divergent_flow = options.divergent_flow
     third_order_terms = options.third_order_terms
     epsilon = options.epsilon
-    DPDC = options.DPDC
+    dpdc = options.DPDC
     dimensionally_split = options.dimensionally_split
 
     @numba.njit(**options.jit_flags)
-    def A(psi):
+    def a_term(psi):
         """ eq. 13 in Smolarkiewicz 1984; eq. 17a in Smolarkiewicz & Margolin 1998 """
         result = ats(*psi, 1) - ats(*psi, 0)
         if infinite_gauge:
@@ -50,7 +50,7 @@ def __make_antidiff(atv, ats, non_unit_g_factor, options, n_dims, last_pass):
[…]
