### 2021-10-28 af4a210
docstring updates, grid.png generation in README
