+++ PyMPDATA/__init__.py
@@ -1,6 +1,10 @@
 """
 Numba-accelerated Pythonic implementation of Multidimensional Positive Definite
 Advection Transport Algorithm (MPDATA) with examples in Python, Julia and Matlab
+
+PyMPDATA uses staggered grid with the following node placement for
+`PyMPDATA.ScalarField` and `PyMPDATA.VectorField` elements:
+![](https://github.com/atmos-cloud-sim-uj/PyMPDATA/releases/download/tip/grid.png)
 """
 # pylint: disable=invalid-name
 from pkg_resources import get_distribution, DistributionNotFound, VersionConflict
