+++ .github/workflows/readme_python.yml
@@ -24,3 +24,11 @@ jobs:
       - run: pip install exdown pytest
       - run: python -c "import exdown; code=exdown.extract('README.md'); f=open('readme.py', 'w'); f.writelines(block.code for block in code if block.syntax=='Python'); f.close()"
       - run: python -We readme.py
+
+      - name: artefacts
+        if: github.ref == 'refs/heads/master'
+        uses: eine/tip@master
+        with:
+          token: ${{ secrets.GITHUB_TOKEN }}
+          files: |
+            grid.png
