+++ .github/workflows/readme_python.yml
@@ -21,7 +21,7 @@ jobs:
         with:
           python-version: 3.9
       - run: pip install -e .
-      - run: pip install exdown pytest
+      - run: pip install exdown pytest matplotlib
       - run: python -c "import exdown; code=exdown.extract('README.md'); f=open('readme.py', 'w'); f.writelines(block.code for block in code if block.syntax=='Python'); f.close()"
       - run: python -We readme.py
 
