+++ README.md
@@ -145,52 +145,51 @@ In most use cases of PyMPDATA, the first thing to do is to instantiate the
 [``Options``](https://atmos-cloud-sim-uj.github.io/PyMPDATA/options.html) class 
 with arguments suiting the problem at hand, e.g.:
 <details>
-<summary>Julia (click to expand)</summary>
+<summary>Julia code (click to expand)</summary>
 
 ```Julia
 using Pkg
 Pkg.add("PyCall")
 using PyCall
 Options = pyimport("PyMPDATA").Options
-options = Options(n_iters=3, infinite_gauge=true, nonoscillatory=true)
+options = Options(n_iters=2)
 ```
 </details>
 <details>
-<summary>Matlab (click to expand)</summary>
+<summary>Matlab code (click to expand)</summary>
 
 ```Matlab
 Options = py.importlib.import_module('PyMPDATA').Options;
 options = Options(pyargs(...
-  'n_iters', 3, ...
-  'infinite_gauge', true, ...
-  'nonoscillatory', true ...
+  'n_iters', 2, ...
 ));
 ```
 </details>
 <details open>
-<summary>Python (click to expand)</summary>
+<summary>Python code (click to expand)</summary>
 
-```python
+```Python
 from PyMPDATA import Options
-options = Options(n_iters=3, infinite_gauge=True, nonoscillatory=True)
+options = Options(n_iters=2)
 ```
 </details>
 
 #### Arakawa-C grid layer and boundary conditions
 
 In PyMPDATA, the solution domain is assumed to extend from the
-first cell's boundary to the last cell's boundary (thus
+first cell's boundary to the last cell's boundary (thus the
 first scalar field value is at ![\[\Delta x/2, \Delta y/2\]](https://render.githubusercontent.com/render/math?math=%5B%5CDelta%20x%2F2%2C%20%5CDelta%20y%2F2%5D)).
 The [``ScalarField``](https://atmos-cloud-sim-uj.github.io/PyMPDATA/scalar_field.html)
-and [``VectorField``](https://atmos-cloud-sim-uj.github.io/PyMPDATA/vector_field.html) classes implementing the
+and [``VectorField``](https://atmos-cloud-sim-uj.github.io/PyMPDATA/vector_field.html) classes implement the
 [Arakawa-C staggered grid](https://en.wikipedia.org/wiki/Arakawa_grids#Arakawa_C-grid) logic
 in which:
-- scalar fields are discretised onto cell-center points,
-- vector fields are discretised onto cell-boundary points.
-The code snippet below generates a schematic of the grid layout in 
-two dimensions:
-<details open>
-<summary>Python (click to expand)</summary>
+- scalar fields are discretised onto cell centres (one value per cell),
+- vector field components are discretised onto cell walls.
+
+The schematic of the employed grid/domain layout in two dimensions is given below
+(with the Python code snippet generating the figure):
+<details>
+<summary>Python code (click to expand)</summary>
 
 ```Python
 import numpy as np
@@ -199,31 +198,36 @@ from matplotlib import pyplot
 dx, dy = .2, .3
 grid = (10, 5)
 
-x, y = np.mgrid[
-    dx / 2 : grid[0] * dx : dx, 
-    dy / 2 : grid[1] * dy : dy
-]
-ux, uy = np.mgrid[
-    0 : (grid[0]+1) * dx : dx, 
-    dy / 2 : grid[1] * dy : dy
-]
-vx, vy = np.mgrid[
-    dx / 2 : grid[0] * dx : dx,
-    0: (grid[1] + 1) * dy : dy
-]
-
[…]
