### 2021-10-30 85c9701
fix Matlab syntax in README.md
