+++ README.md
@@ -278,14 +278,26 @@ Periodic = py.importlib.import_module('PyMPDATA.boundary_conditions').Periodic;
 
 nx = int32(24);
 ny = int32(24);
+  
+Cx = -.5;
+Cy = -.25;
+
+[xi, yi] = meshgrid(double(0:1:nx-1), double(0:1:ny-1));
+
 halo = options.n_halo;
 advectee = ScalarField(pyargs(...
-    'data', py.numpy.zeros(int32([nx ny])), ... 
+    'data', py.numpy.array(exp( ...
+        -(xi+.5-double(nx)/2).^2 / (2*(double(nx)/10)^2) ...
+        -(yi+.5-double(ny)/2).^2 / (2*(double(ny)/10)^2) ...
+    )), ... 
     'halo', halo, ...
     'boundary_conditions', py.tuple({Periodic(), Periodic()}) ...
 ));
 advector = VectorField(pyargs(...
-    'data', py.tuple({py.numpy.zeros(int32([nx+1 ny])), py.numpy.zeros(int32([nx ny+1]))}), ...
+    'data', py.tuple({ ...
+        Cx * py.numpy.ones(int32([nx+1 ny])), ... 
+        Cy * py.numpy.ones(int32([nx ny+1])) ...
+     }), ...
     'halo', halo, ...
     'boundary_conditions', py.tuple({Periodic(), Periodic()}) ...
 ));
@@ -354,7 +366,7 @@ stepper = Stepper(options=options, n_dims=2)
 ```Matlab
 Stepper = py.importlib.import_module('PyMPDATA').Stepper;
 
-steper = Stepper(pyargs(...
+stepper = Stepper(pyargs(...
   'options', options, ...
   'n_dims', int32(2) ...
 ));
