+++ README.md
@@ -255,14 +255,19 @@ VectorField = pyimport("PyMPDATA").VectorField
 Periodic = pyimport("PyMPDATA.boundary_conditions").Periodic
 
 nx, ny = 24, 24
+Cx, Cy = -.5, -.25
+idx = CartesianIndices((nx, ny))
 halo = options.n_halo
 advectee = ScalarField(
-    data=zeros((nx, ny)), 
+    data=exp.(
+        -(getindex.(idx, 1) .- .5 .- nx/2).^2 / (2*(nx/10)^2) 
+        -(getindex.(idx, 2) .- .5 .- ny/2).^2 / (2*(ny/10)^2)
+    ),  
     halo=halo, 
     boundary_conditions=(Periodic(), Periodic())
 )
 advector = VectorField(
-    data=(zeros((nx+1, ny)), zeros((nx, ny+1))),
+    data=(fill(Cx, (nx+1, ny)), fill(Cy, (nx, ny+1))),
     halo=halo,
     boundary_conditions=(Periodic(), Periodic())    
 )
