### 2021-11-01 c78ea00
adding code samples and figures
