+++ paper/paper.md
@@ -80,7 +80,6 @@ The appendices of the present paper include ``Python``, ``Julia`` and ``Matlab``
   depicted in Fig. (\autoref{fig}) and based on Fig. 5 from @Arabas_et_al_2014.
 
 \begin{figure}[t]
-    \centering
     \includegraphics[width=0.5\textwidth]{readme_gauss_0-crop}
     \includegraphics[width=0.5\textwidth]{readme_gauss-crop}
     \caption{Visualisation of the initial condition (left) and simulation state after 75 timesteps (right)
