### 2021-11-01 #3855 @slayoo
(updated the paper.md with a figure and sample code snippets in Python, Julia and Matlab)
