### 2021-11-04 #3855 @slayoo
Hello, @mjsottile,
At the bottom of the PyMPDATA README file, we have listed several related projects (https://github.com/atmos-cloud-sim-uj/PyMPDATA#other-open-source-mpdata-implementations). Among contributors to these projects, and excluding folks we have co-authored with, there are the following Github users - perhaps potential reviewers:
- hilaryweller0 (https://github.com/AtmosFOAM/AtmosFOAM/commit/947b192f69d973ea4a7cfab077eb5c6c6fa8b0cf)
- dmikushin (https://github.com/apc-llc/mpdata/graphs/contributors)
- anilzen, olekravchenko (https://github.com/olekravchenko/centpy/graphs/contributors)
- ThanasisMattas (https://github.com/ThanasisMattas/mattflow/graphs/contributors)
- stefanmeili (https://github.com/stefanmeili/FastFD/graphs/contributors)
- https://github.com/orgs/clawpack/people

Browsing through Github topics, some other projects can be found and filtered by language, e.g.:
- https://github.com/topics/advection-diffusion?l=Python
- https://github.com/topics/advection?l=Matlab
- https://github.com/topics/advection-diffusion?l=Julia

HTH, thank you,
Sylwester
