### 2021-11-07 #3896 @whedon
[REVIEW]: PyMPDATA v1: Numba-accelerated Pythonic implementation of MPDATA with examples in Python, Julia and Matlab
**Submitting author:** @slayoo (<a href="http://orcid.org/0000-0003-2361-0082">Sylwester Arabas</a>)
**Repository:** https://github.com/atmos-cloud-sim-uj/PyMPDATA
**Branch with paper.md** (empty if default branch):
**Version:** v1
