### 2021-11-07 #3896 @slayoo
it's puzzling as the above "INVALID DOIs" seem correct, at leat both of the below URLs redirect correctly to the journal site:
- https://doi.org/10.1175/1520-0450(1968)007%3C0160:AOFDMA%3E2.0.CO;2
- https://doi.org/10.1175/1520-0493(1989)117%3C0102:TDSLTW%3E2.0.CO;2
