### 2021-11-07 #3896 @danielskatz
Please make sure the DOIs are in doi entries in the bib file rather than url entries, and that they don't include the https://doi.org part
