### 2021-11-07 #3896 @danielskatz
And perhaps change the characters that are being encoded for example:

10.1175/1520-0493(1989)117<0102:TDSLTW>2.0.CO;2
