+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Molenkamp_1968,
-  doi = {10.1175/1520-0450(1968)007\%3C0160:AOFDMA\%3E2.0.CO;2},
+  doi = {10.1175/1520-0450(1968)007<0160:AOFDMA>2.0.CO;2},
   title = {Accuracy of Finite-Difference Methods Applied to the Advection Equation},
   journal = {J. Appl. Meteorol. Climatol.},
   year = {1968},
@@ -49,7 +49,7 @@
   year = {1989},
   volume = {117},
   title = {Two-Dimensional Semi-{L}agrangian Transport with Shape-Preserving Interpolation },
-  doi = {10.1175/1520-0493(1989)117\%3C0102:TDSLTW\%3E2.0.CO;2}
+  doi = {10.1175/1520-0493(1989)117<0102:TDSLTW>2.0.CO;2}
 }
 
 @article{Smolarkiewicz_and_Grabowski_1990,
