### 2021-11-08 #3896 @slayoo
thank you @danielskatz, indeed the URL escapes in the bib file were the cause - good job on the checker side!
