+++ paper/paper.bib
@@ -119,7 +119,7 @@
     journal = {Int. J. Numer. Methods Fluids},
     volume={50},
     year = {2006},
-    doi = {doi:10.1002/fld.1071}
+    doi = {10.1002/fld.1071}
 }
 
 @phdthesis{Hill_2010,
