### 2021-12-14 8dad5e3
adding -We to python for pdoc invocation to flag broken references in docstrings as CI build failures
