+++ .github/workflows/pdoc.yml
@@ -20,7 +20,7 @@ jobs:
         run: |
           pip3 install pdoc3
           pip install -e .
-          pdoc --html PyMPDATA
+          python -We -m pdoc --html PyMPDATA
       - name: Deploy
         if: github.ref == 'refs/heads/master'
         uses: JamesIves/github-pages-deploy-action@4.1.1
