+++ PyMPDATA/__init__.py
@@ -3,7 +3,8 @@ Numba-accelerated Pythonic implementation of Multidimensional Positive Definite
 Advection Transport Algorithm (MPDATA) with examples in Python, Julia and Matlab
 
 PyMPDATA uses staggered grid with the following node placement for
-`PyMPDATA.ScalarField` and `PyMPDATA.VectorField` elements:
+`PyMPDATA.scalar_field.ScalarField` and
+`PyMPDATA.vector_field.VectorField` elements:
 ![](https://github.com/atmos-cloud-sim-uj/PyMPDATA/releases/download/tip/readme_grid.png)
 """
 # pylint: disable=invalid-name
