### 2021-12-14 ab864ff
more docstring refs fixes
