+++ PyMPDATA/boundary_conditions/extrapolated.py
@@ -7,7 +7,8 @@ from PyMPDATA.impl.enumerations import META_AND_DATA_META, META_AND_DATA_DATA
 
 class Extrapolated:
     """ class which instances are to be passed in boundary_conditions tuple to the
-        `PyMPDATA.ScalarField` and `PyMPDATA.VectorField` __init__ methods """
+        `PyMPDATA.scalar_field.ScalarField` and
+        `PyMPDATA.vector_field.VectorField` __init__ methods """
     def __init__(self, dim=INNER, eps=1e-10):
         self.eps = eps
         self.dim = dim
