### 2021-12-14 2b0cb52
adding docstring for _HashableDict
