### 2021-12-14 ce81533
deprivitizing HashableDict
