+++ .github/workflows/readme_python.yml
@@ -21,8 +21,8 @@ jobs:
         with:
           python-version: 3.9
       - run: pip install -e .
-      - run: pip install exdown pytest matplotlib
-      - run: python -c "import exdown; code=exdown.extract('README.md'); f=open('readme.py', 'w'); f.writelines(block.code for block in code if block.syntax=='Python'); f.close()"
+      - run: pip install pytest-codeblocks pytest matplotlib
+      - run: python -c "import pytest-codeblocks; code=pytest-codeblocks.extract_from_file('README.md'); f=open('readme.py', 'w'); f.writelines(block.code for block in code if block.syntax=='Python'); f.close()"
       - run: python -We readme.py
 
       - name: artefacts
