+++ .github/workflows/readme_python.yml
@@ -22,7 +22,7 @@ jobs:
           python-version: 3.9
       - run: pip install -e .
       - run: pip install pytest-codeblocks pytest matplotlib
-      - run: python -c "import pytest-codeblocks; code=pytest-codeblocks.extract_from_file('README.md'); f=open('readme.py', 'w'); f.writelines(block.code for block in code if block.syntax=='Python'); f.close()"
+      - run: python -c "import pytest_codeblocks; code=pytest_codeblocks.extract_from_file('README.md'); f=open('readme.py', 'w'); f.writelines(block.code for block in code if block.syntax=='Python'); f.close()"
       - run: python -We readme.py
 
       - name: artefacts
