### 2021-12-31 #3896 @dmikushin
Hi @slayoo , is there a performance comparison between Numba-based PyMPDATA and any other native code implementations of your choice? That would be really interesting to see!
