### 2022-01-03 #3896 @slayoo
@dmikushin Thanks for feedback. Very good point on including a performance comparison with other implementations,

We have so far:
- looked at wall times for these tests which have direct counterparts in libmpdata++ (Github Actions used in both projects - times on similar virtual machines are available for each CI build, but these include compilation times in the case of PyMPDATA, hence not straightforward to analyse "as is", but that would be the first candidate to start the analysis from);
- analysed execution times of various MPDATA varians versus upwind times for PyMPDATA and corroborated it with such wall-time ratios from literature, see Table 2 here: https://arxiv.org/pdf/2011.14726.pdf

Comparing with libmpdata++ times is certainly doable (February/March?).
Any other suggestion on what to compare with?
