### 2022-01-14 bf4f13e
bump numba req for CI; add P3.10 to CI
