+++ .github/workflows/pytest.yml
@@ -17,7 +17,7 @@ jobs:
     strategy:
       matrix:
         platform: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: [3.7, 3.8 , 3.9]
+        python-version: [3.7, 3.8 , 3.9, 3.10]
     runs-on: ${{ matrix.platform }}
     steps:
       - uses: actions/checkout@v2
