### 2022-01-14 ccaca69
indicate incompatibility with Numba 0.55 in setup.py
