### 2022-01-15 a121c17
pass all null fields as arguments to njitted code (compatibility with Numba 0.55)
