+++ PyMPDATA/impl/formulae_laplacian.py
@@ -26,11 +26,10 @@ def make_laplacian(non_unit_g_factor: bool, options: Options, traversals: Traver
             for i in range(MAX_DIM_NUM)
         )
 
-        null_vecfield, null_vecfield_bc = traversals.null_vector_field.impl
-        null_scalarfield, null_scalarfield_bc = traversals.null_scalar_field.impl
-
         @numba.njit(**options.jit_flags)
-        def apply(advector, advectee, advectee_bc):
+        def apply(null_impl, advector, advectee, advectee_bc):
+            null_vecfield, null_vecfield_bc = null_impl.vector
+            null_scalarfield, null_scalarfield_bc = null_impl.scalar
             return apply_vector(*formulae_laplacian,
                                 *advector,
                                 *advectee, *advectee_bc,
