+++ PyMPDATA/impl/traversals_vector.py
@@ -74,14 +74,14 @@ def _make_apply_vector(*, indexers, jit_flags, halo, n_dims, n_threads,
                                  arg1s_bc_o, arg1s_bc_m, arg1s_bc_i)
             boundary_cond_vector(thread_id, arg2v_meta, arg2v_data_o, arg2v_data_m, arg2v_data_i,
                                  arg2v_bc_o, arg2v_bc_m, arg2v_bc_i)
-            boundary_cond_scalar(thread_id, arg3s_meta, arg3s_data,
-                                 arg3s_bc_o, arg3s_bc_m, arg3s_bc_i)
+            # boundary_cond_scalar(thread_id, arg3s_meta, arg3s_data,
+            #                      arg3s_bc_o, arg3s_bc_m, arg3s_bc_i)
         if not arg1s_meta[META_HALO_VALID]:
             arg1s_meta[META_HALO_VALID] = True
         if not arg2v_meta[META_HALO_VALID]:
             arg2v_meta[META_HALO_VALID] = True
-        if not arg3s_meta[META_HALO_VALID]:
-            arg3s_meta[META_HALO_VALID] = True
+        # if not arg3s_meta[META_HALO_VALID]:
+        #     arg3s_meta[META_HALO_VALID] = True
 
         for thread_id in range(1) if n_threads == 1 else numba.prange(n_threads):
             apply_vector_impl(
