+++ PyMPDATA/impl/formulae_flux.py
@@ -8,7 +8,6 @@ def make_flux_first_pass(options, traversals):
     """ returns njit-ted function for use with given traversals """
     idx = traversals.indexers[traversals.n_dims]
     apply_vector = traversals.apply_vector()
-    null_scalarfield, null_bc = traversals.null_scalar_field.impl
 
     formulae_flux_first_pass = tuple(
         __make_flux(
@@ -23,7 +22,8 @@ def make_flux_first_pass(options, traversals):
     )
 
     @numba.njit(**options.jit_flags)
-    def apply(vectmp_a, advector, advectee, advectee_bc, vec_bc):
+    def apply(null_impl, vectmp_a, advector, advectee, advectee_bc, vec_bc):
+        null_scalarfield, null_bc = null_impl.scalar
         return apply_vector(*formulae_flux_first_pass,
                             *vectmp_a,
                             *advectee, *advectee_bc,
@@ -49,10 +49,9 @@ def make_flux_subsequent(options, traversals):
         for i in range(MAX_DIM_NUM)
     )
 
-    null_scalarfield, null_bc = traversals.null_scalar_field.impl
-
     @numba.njit(**options.jit_flags)
-    def apply(flux, psi, psi_bc, g_c_corr, vec_bc):
+    def apply(null_impl, flux, psi, psi_bc, g_c_corr, vec_bc):
+        null_scalarfield, null_bc = null_impl.scalar
         return apply_vector(*formulae_flux_subsequent,
                             *flux,
                             *psi, *psi_bc,

[…]
