### 2022-01-15 383570a
simplify passing bc by using namedtuples
