### 2022-01-15 e4e4ea7
fixes in optimised null formulae
