### 2022-01-15 a402d8e
bump numpy req, resolve namedtuple docstring issue
