+++ tests/unit_tests/test_formulae_upwind.py
@@ -4,7 +4,7 @@ from numba.core.errors import NumbaExperimentalFeatureWarning
 import numpy as np
 from PyMPDATA.impl.formulae_upwind import make_upwind
 from PyMPDATA.impl.traversals import Traversals
-from PyMPDATA.impl.meta import Impl
+from PyMPDATA.impl.meta import _Impl
 from PyMPDATA.impl.enumerations import IMPL_META_AND_DATA, IMPL_BC
 from PyMPDATA import Options, ScalarField, VectorField
 from PyMPDATA.boundary_conditions import Periodic
@@ -40,9 +40,9 @@ def test_formulae_upwind():
         warnings.simplefilter('ignore', category=NumbaExperimentalFeatureWarning)
         upwind(
             traversals.null_impl,
-            Impl(field=psi_impl[IMPL_META_AND_DATA], bc=psi_impl[IMPL_BC]),
-            Impl(field=flux_impl[IMPL_META_AND_DATA], bc=flux_impl[IMPL_BC]),
-            Impl(
+            _Impl(field=psi_impl[IMPL_META_AND_DATA], bc=psi_impl[IMPL_BC]),
+            _Impl(field=flux_impl[IMPL_META_AND_DATA], bc=flux_impl[IMPL_BC]),
+            _Impl(
                field=traversals.null_impl.scalar[IMPL_META_AND_DATA],
                bc=traversals.null_impl.scalar[IMPL_BC]
             )
