### 2022-01-16 b4e658a
fix pip call order (i.e. add another call)
