+++ .github/workflows/pytest.yml
@@ -32,6 +32,8 @@ jobs:
           python -We -c "import PyMPDATA"
       - run: |
           pip install -r test-time-requirements.txt
+      - run: |
+          pip install -e .
       # https://github.com/numba/numba/issues/6350#issuecomment-728174860
       - if: matrix.platform == 'ubuntu-latest'
         env:
