### 2022-01-17 bb986f8
fix pip order for codecov workflow
