+++ .github/workflows/codecov.yml
@@ -12,8 +12,8 @@ jobs:
         python-version: 3.9
     - name: Generate coverage report
       run: |
-        pip install -e .
         pip install -r test-time-requirements.txt
+        pip install -e .
         pip install pytest-cov
         pytest tests/unit_tests --cov-report=xml --cov=PyMPDATA
     - name: Upload coverage to Codecov
