### 2022-01-17 9846130
add skipif("numba.config.DISABLE_JIT") to test_shared_advector
