### 2022-01-17 a1f6b91
use lru_cache in polar make_vector (fixes the type unification problem with Numba as gives the same compiled func address)
