### 2022-01-17 5ac5924
saner handling of boundary_conditions when cloning
