### 2022-01-30 bc706cb
bump numba req
