### 2022-02-10 #3896 @arfon
Hi all, I'm going to take over here as editor and help @mjsottile out a little.

@dmikushin, @olekravchenko – are you still able to review this submission for JOSS? If not, I will proceed to identify new reviewer candidates.
