+++ .github/workflows/pylint.yml
@@ -2,9 +2,9 @@ name: pylint
 
 on:
   push:
-    branches: [ master ]
+    branches: [ main ]
   pull_request:
-    branches: [ master ]
+    branches: [ main ]
 
 jobs:
   build:
