### 2022-02-23 283c6e8
trying with pywinpty!=2.0.2
