+++ appveyor.yml
@@ -27,6 +27,9 @@ install:
   - ps: |
       iex "$env:PYTHON -m pip install --upgrade pip"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
+
+      iex "$env:PYTHON -m pip install pywinpty!=2.0.2"
+      if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
       
       iex "$env:PYTHON -m pip install -r test-time-requirements.txt"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
