### 2022-02-23 ce5568f
trying with pywinpty 1.1.6
