+++ appveyor.yml
@@ -28,7 +28,7 @@ install:
       iex "$env:PYTHON -m pip install --upgrade pip"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
 
-      iex "$env:PYTHON -m pip install pywinpty!=2.0.2"
+      iex "$env:PYTHON -m pip install pywinpty==1.1.6"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
       
       iex "$env:PYTHON -m pip install -r test-time-requirements.txt"
