+++ appveyor.yml
@@ -28,9 +28,11 @@ install:
       iex "$env:PYTHON -m pip install --upgrade pip"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
 
-      iex "$env:PYTHON -m pip install pywinpty==1.1.6"
+  - ps: |
+      iex "$env:PYTHON -m pip install pywinpty==1.1.0"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
-      
+
+  - ps: |
       iex "$env:PYTHON -m pip install -r test-time-requirements.txt"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
 
