### 2022-02-23 49d74a9
try installing rustc and cargo
