+++ appveyor.yml
@@ -26,19 +26,8 @@ install:
         $env:PATH += ';C:\msys64\mingw64\bin'
       }
   - ps: |
-      iex "$env:PYTHON -m pip install --upgrade pip"
+      iex "$env:PYTHON -m pip install pip==21.3.1"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
-
-  - ps: |
-       curl -sSf -o rustup-init.exe https://win.rustup.rs/
-       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
-  - ps: |
-       rustup-init.exe -y --default-host %TARGET% --default-toolchain %RUST_VERSION%
-       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
-  - ps: |
-       set PATH=%PATH%;C:\Users\appveyor\.cargo\bin
-       rustc -Vv
-       cargo -V
   - ps: |
       iex "$env:PYTHON -m pip install -r test-time-requirements.txt"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
