### 2022-02-28 #3896 @arfon
:wave: @Chiil and @wdeconinck – would either of you be willing to review this submission for JOSS? The submission under consideration is [_PyMPDATA v1: Numba-accelerated Pythonic implementation of MPDATA with examples in Python, Julia and Matlab_](https://github.com/atmos-cloud-sim-uj/PyMPDATA).
