### 2022-02-28 #3896 @arfon
> Sure, I can review it.

Amazing, thank you! You can generate a review checklist by typing the following which will make you a checklist to work through:

Our reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.
