### 2022-02-28 #3896 @arfon
> I am happy to review as well.

Thank you very much @wdeconinck! You can generate a review checklist by typing the following which will make you a checklist to work through:

Our reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.
