+++ paper/paper.bib
@@ -207,10 +207,10 @@
 
 @article{Olesik_et_al_2021,
   author = {Olesik, M. and Banaśkiewicz, J. and Bartman, P. and Baumgartner, M. and Unterstrasser, S. and Arabas, S.},
-  title = {On numerical broadening of particle size spectra: a condensational growth study using PyMPDATA},
-  journal = {Geosci. Model Dev. Discuss.},
+  title = {On numerical broadening of particle-size spectra: a condensational growth study using PyMPDATA 1.0},
+  journal = {Geosci. Model Dev.},
   doi = {10.5194/gmd-2020-404},
-  year = {2021}
+  year = {2022}
 }
 
 @article{Bartman_et_al_2021,
