### 2022-04-18 #3896 @slayoo
@arfon the check-references behaviour has apparently changed recently wrt handling of "<" and ">" characters in the DOIs - the above four invalid DOIs were not indicated as invalid up above: https://github.com/openjournals/joss-reviews/issues/3896#issuecomment-962827845
Was it intentional? Thanks,
Sylwester
