### 2022-04-23 #3896 @slayoo
Thank you @wdeconinck for the feedback!
Our plan is to add to the paper an analysis comparing libmpdata++ and PyMPDATA execution times for 1-2 examples that are already implemented in both packages. For both packages, we will check how the time scales with the number of threads, and for PyMPDATA we will report times with Numba enabled and without JIT. Please stay tuned for an update.
