### 2022-05-05 #3896 @Chiil
Furthermore, I have not checked the "State of the field" box, because the paper does not contain a comparison against other packages that can solve advection diffusion equations. It does show a nice overview though in the `README.md` of other packages that can do the same. A small comparison highlighting the benefits of PyMPDATA could improve the paper.
