### 2022-05-05 #3896 @Chiil
And for curiousity, what are the prospect of running this package on GPUs, given that `numba` is also improving its GPU support.
