### 2022-06-18 1e9362a
addressing pylint suggestions
