### 2022-06-18 4b1e4b7
addressing one more pylint suggestion
