+++ tests/smoke_tests/kinematic_2d/test_single_timestep.py
@@ -17,8 +17,8 @@ def stream_function(x_01, z_01):
     return - RHOD_W_MAX * x_span / np.pi * np.sin(np.pi * z_01) * np.cos(2 * np.pi * x_01)
 
 
-def rhod_of_z(z):
-    return 1 - z * 1e-4
+def rhod_of_z(arg):
+    return 1 - arg * 1e-4
 
 RHOD = np.repeat(
     rhod_of_z(
