### 2022-07-01 #3896 @slayoo
> I have tested all examples, and the only one that did not work (and I hope the problem is not on my Julia installation side) is the Julia one. I have tested the code from the Appendix of `paper.md` and got:
>
> ```
> julia> using PyCall
>
> julia> Options = pyimport("PyMPDATA").Options
> Assertion failed: (findOption(Name) == Values.size() && "Option already exists!"), function addLiteralOption, file /Users/ci/miniconda3/conda-bld/llvmdev_1601977477900/work/include/llvm/Support/CommandLine.h, line 853.
>
> signal (6): Abort trap: 6
> in expression starting at REPL[2]:1
> __pthread_kill at /usr/lib/system/libsystem_kernel.dylib (unknown line)
> pthread_kill at /usr/lib/system/libsystem_pthread.dylib (unknown line)
> abort at /usr/lib/system/libsystem_c.dylib (unknown line)
> __assert_rtn at /usr/lib/system/libsystem_c.dylib (unknown line)
> _ZN4llvm2cl6parserIPFPNS_12FunctionPassEvEE16addLiteralOptionIS5_EEvNS_9StringRefERKT_S8_ at /usr/local/lib/python3.9/site-packages/llvmlite/binding/libllvmlite.dylib (unknown line)
> _ZN4llvm2cl3optIPFPNS_12FunctionPassEvELb0ENS_18RegisterPassParserINS_16RegisterRegAllocEEEEC2IJA9_cNS0_12OptionHiddenENS0_11initializerIS5_EENS0_4descEEEEDpRKT_ at /usr/local/lib/python3.9/site-packages/llvmlite/binding/libllvmlite.dylib (unknown line)
> _GLOBAL__sub_I_TargetPassConfig.cpp at /usr/local/lib/python3.9/site-packages/llvmlite/binding/libllvmlite.dylib (unknown line)
> Allocations: 716693 (Pool: 716197; Big: 496); GC: 1
> Abort trap: 6
> ```

IIUC, this seem to be a problem with Julia and Numba being linked to different versions of LLVM. There is little we can do at the level of PyMPDATA. The problem has been reported in numerous places, with no solution provided in most cases besides just upgrading/downgrading packages (on Github Actions, we do test the Julia code snippet on each merge to main, and there is no trouble on `ubuntu-latest`):
- https://github.com/JuliaPy/PyCall.jl/issues/861
- https://github.com/JuliaPy/PyCall.jl/issues/976
- https://github.com/e2nIEE/pandapower/issues/1217
- https://github.com/JuliaPy/pyjulia/issues/443
- https://github.com/JuliaPy/PyCall.jl/issues/903
