### 2022-07-01 #3896 @slayoo
> I do get a lot of deprecation warnings though in some of the examples. For instance the `advection_diffusion_1d` notebook give me a lot of warnings like
>
> ```
> /var/folders/4f/nykt4nhx4kn12sz633_msth40000gn/T/ipykernel_45402/528486334.py:52: DeprecationWarning: `np.int` is a deprecated alias for the builtin `int`. To silence this warning, use `int` by itself. Doing this will not modify any behavior and is safe. When replacing `np.int`, you may wish to use e.g. `np.int64` or `np.int32` to specify the precision. If you wish to review your current use, check the release note link for additional information.
> Deprecated in NumPy 1.20; for more details and guidance: https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
>   steps = np.arange(nt + 1, dtype=np.int).reshape(-1, 1) * steps_per_iter
> ```

Fixed here: https://github.com/atmos-cloud-sim-uj/PyMPDATA-examples/commit/763c004a890088ea662fe591ced5e8feaad5df19
