### 2022-07-04 c4dfab1
fix documentation for timers (wall-time -> CPU time)
