### 2022-07-05 #3896 @slayoo
Just a quick update regarding the performance analysis. We are done with JIT enabled vs. disabled + threading analysis, still working on comparison with libmpdata++. Here's a draft of the figure based on a 3D advection-only example from Smolarkiewicz 1984, and depicting that:
- disabling Numba JIT causes an over 1000-fold slowdown in PyMPDATA;
- there is reasonable speedup measurable from using multiple threads in PyMPDATA;
- switching from dynamic grid (grid size decided at runtime) vs. static grid (grid size decided at compile time) gives noticeable and consistent speedup, although hardly striking when compared with the JIT-disabled timings on one plot.
![image](https://user-images.githubusercontent.com/89685/177293196-22cd3e60-eb09-4886-82ad-ac87745944b5.png)

Please stay tuned a few days more for addition of libmpdata++ datapoints to the figure and for a PR merging the notebook covering this analysis into the PyMPDATA-examples repo, and a relevant additions to the JOSS manuscript.
