### 2022-07-05 #3896 @wdeconinck
That is very nice. I guess if you want to make the point of comparing the three PyMPDATA results, you would create a second picture with zoomed vertical axis.
