### 2022-07-06 39fad9c
ignore DeprecationWarning for fastcore import
