### 2022-07-12 #3896 @slayoo
Here's an updated plot with libmpdata++ timings added, and with the JIT-disabled timing shown in a separate subplot. It shows that, in this particular case (3D problem, 32x32x32 grid, 3 MPDATA iterations), PyMPDATA performs a even better than libmpdata++, and over 1000-times faster than with Numba JIT disabled.
![Screenshot from 2022-07-13 00-24-07](https://user-images.githubusercontent.com/89685/178606739-793fdb3f-21ae-4703-aa72-56a06a01bcdb.png)

Wall-times taken as minimal values over 5 repetitions ensuring that JIT-compilation times, thread scheduling or similar effects are filtered out.

A notebook orchestrating the analysis and plot generation plus C++ code for libmpdata++ timing is on its way to PyMPDATA-examples repo: https://github.com/atmos-cloud-sim-uj/PyMPDATA-examples/pull/64

paper text update coming very soon
