### 2022-07-13 77f2541
add section+figure on performance analysis vs. libmpdata++
