+++ paper/paper.bib
@@ -214,8 +214,7 @@
 }
 
 @article{Bartman_et_al_2021,
-  author = {Bartman, P. and Bulenok, O. and Górski, K. and Jaruga, A. and Łazarski G. and Olesik, M. and
-Piasecki, B. and Singer, C.E. and Talar, A. and Arabas, S.},
+  author = {Bartman, P. and Bulenok, O. and Górski, K. and Jaruga, A. and Łazarski G. and Olesik, M. and Piasecki, B. and Singer, C.E. and Talar, A. and Arabas, S.},
   title = {PySDM v1: particle-based cloud modelling package for warm-rain microphysics and aqueous chemistry},
   year = {2021},
   volume = {under review in JOSS, e-print at arXiv:2103.17238},
