### 2022-07-13 6b99f93
updated performance figure (removed init time for PyMPDATA) + clarification in the text
