### 2022-07-13 e26106b
more comments on performance
