### 2022-07-13 19974e9
add section+figure on performance analysis vs. libmpdata++
