### 2022-07-13 #3896 @slayoo
@wdeconinck @Chiil, in the above updated pdf, the manuscript is extended with a 1.5-page long "Performance" section. There are two new plots composed into Fig. 2 which present analysis of scaling with the number of threads and scaling with the domain size. The plots include wall-time measurements for PyMPDATA with Numba JIT disabled and enabled, as well as measurements for libmpdata++. Comments welcome, thanks.
