### 2022-07-18 #3896 @wdeconinck
@slayoo That looks great, it shows off very nicely to be on par with the classic HPC languages or even better, with the added benefits of Python.
One typo in line 130: built-in instead of build-in.
