### 2022-07-28 3ba899e
rephrase and fix typo around timing and halo-filling passages
