### 2022-07-29 #3896 @slayoo
Thank you, @wdeconinck.
Typo fixed in https://github.com/atmos-cloud-sim-uj/PyMPDATA/commit/3ba899ecfbd2f99d1426e534585fba57e532eaea
The performance-analysis notebook is now merged (https://github.com/atmos-cloud-sim-uj/PyMPDATA-examples/pull/64) and is executed as a part of CI including compilation, execution and timing of libmpdata++ code (but only on Linux).
New v1.0.3 package version also released and on its way to PyPI.org.

@wdeconinck, @Chiil, please confirm, by marking relevant checkboxes in the reviewers' checklist, if the added material comparing performance with libmpdata++ is sufficient? Thank you
