### 2022-08-02 3cc6271
bump numba requirement
