### 2022-08-12 #3896 @arfon
> @wdeconinck, @Chiil, please confirm, by marking relevant checkboxes in the reviewers' checklist, if the added material comparing performance with libmpdata++ is sufficient? Thank you

Yes, please let us know if you're satisfied with these changes.
