+++ setup.py
@@ -23,7 +23,7 @@ setup(
     setup_requires=['setuptools_scm'],
     install_requires=[
         'numba' + ('==0.56.0' if 'CI' in os.environ else ''),
-        'numpy' + ('==1.22.4' if 'CI' in os.environ else ''),
+        'numpy' + ('==1.21.6' if 'CI' in os.environ else ''),
         'pystrict'
     ],
     author='https://github.com/atmos-cloud-sim-uj/PyMPDATA/graphs/contributors',
