### 2022-08-29 #3896 @slayoo
@Chiil (CC: @arfon), please let us know if the added section with performance comparison against libmpdata++ is be enough to pass the missing checkbox item in your review?

Your checklist is here:
https://github.com/openjournals/joss-reviews/issues/3896#issuecomment-1093951893

(with the missing checkbox here:
![image](https://user-images.githubusercontent.com/89685/187143227-c68b9c6e-0001-404d-909d-56c061c36340.png)
)

The added section is in the most recent pdf preview:
https://raw.githubusercontent.com/openjournals/joss-papers/joss.03896/joss.03896/10.21105.joss.03896.pdf

Thanks!
