+++ .github/workflows/tests+pypi.yml
@@ -73,7 +73,7 @@ jobs:
           repository_url: https://test.pypi.org/legacy/
 
       - if: startsWith(github.ref, 'refs/tags')
-        uses: pypa/gh-action-pypi-publish@master
+        uses: pypa/gh-action-pypi-publish@unstable/v1
         with:
           password: ${{ secrets.PYPI_API_TOKEN }}
 
