### 2022-09-04 #3896 @arfon
@slayoo – this line in your BibTeX is tripping up the XML validation: https://github.com/atmos-cloud-sim-uj/PyMPDATA/blob/main/paper/paper.bib#L220

Could you change it to be a valid preprint BibTeX entry?
