### 2022-09-04 1ad49aa
bibliography cleanup in JOSS paper
