+++ paper/paper.bib
@@ -1,19 +1,19 @@
 @article{Molenkamp_1968,
-  doi = {10.1175/1520-0450(1968)007<0160:AOFDMA>2.0.CO;2},
-  title = {Accuracy of Finite-Difference Methods Applied to the Advection Equation},
-  journal = {J. Appl. Meteorol. Climatol.},
-  year = {1968},
-  volume = {7},
-  author = {Molenkamp, C.R.}
+    doi = {10.1175/1520-0450(1968)007<0160:AOFDMA>2.0.CO;2},
+    title = {Accuracy of Finite-Difference Methods Applied to the Advection Equation},
+    journal = {J. Appl. Meteorol. Climatol.},
+    year = {1968},
+    volume = {7},
+    author = {Molenkamp, C.R.}
 }
 
 @article{Smolarkiewicz_1983,
-  title = {A simple positive definite advection scheme with small implicit diffusion},
-  author = {Smolarkiewicz, P.K.},
-  volume = {111},
-  year = {1983},
-  journal = {Mon. Weather Rev.},
-  doi = {10.1175/1520-0493(1983)111<0479:ASPDAS>2.0.CO;2}
+    title = {A simple positive definite advection scheme with small implicit diffusion},
+    author = {Smolarkiewicz, P.K.},
+    volume = {111},
+    year = {1983},
+    journal = {Mon. Weather Rev.},
+    doi = {10.1175/1520-0493(1983)111<0479:ASPDAS>2.0.CO;2}
 }
 
 @article{Smolarkiewicz_1984,
@@ -35,7 +35,7 @@
 }
 
 @inproceedings{Beason_Margolin_1988,
-  title = "DPDC (double-pass donor cell): A second-order monotone scheme for advection",
+    title = "DPDC (double-pass donor cell): A second-order monotone scheme for advection",
     year = "1988",
     booktitle={Fifth Nuclear Code Developers' Conference},
     note = {{LLNL} report UCRL-99731},
@@ -44,12 +44,12 @@
 }
 
 @article{Williamson_and_Rasch_1989,
-  author = {Williamson, D.L. and Rasch, P.J.},
-  journal = {Mon. Weather Rev.},
-  year = {1989},
-  volume = {117},
-  title = {Two-Dimensional Semi-{L}agrangian Transport with Shape-Preserving Interpolation },
-  doi = {10.1175/1520-0493(1989)117<0102:TDSLTW>2.0.CO;2}
+    author = {Williamson, D.L. and Rasch, P.J.},
+    journal = {Mon. Weather Rev.},
+    year = {1989},
+    volume = {117},
+    title = {Two-Dimensional Semi-{L}agrangian Transport with Shape-Preserving Interpolation },
+    doi = {10.1175/1520-0493(1989)117<0102:TDSLTW>2.0.CO;2}
 }
 
 @article{Smolarkiewicz_and_Grabowski_1990,
@@ -58,7 +58,7 @@
     year = {1990},
     doi = {10.1016/0021-9991(90)90105-A},
     journal = {J. Comp. Phys.},
-    volume={86}
+    volume = {86}
 }
 
 @article{Smolarkiewicz_and_Margolin_1993,
@@ -71,11 +71,11 @@
 }
 
 @book{Morton_1996,
-  year = {1996},
-  author = {Morton, K.W.},
-  title = {Numerical Solution of Convection-Diffusion Problems},
-  doi = {10.1201/9780203711194},
-  publisher = {CRC Press}
+    year = {1996},
+    author = {Morton, K.W.},
+    title = {Numerical Solution of Convection-Diffusion Problems},
+    doi = {10.1201/9780203711194},
+    publisher = {CRC Press}
 }
 
 @article{Smolarkiewicz_and_Margolin_1998_JCP,
@@ -92,7 +92,9 @@
[…]
