### 2022-09-04 #3896 @slayoo
@arfon, fixed here: https://github.com/atmos-cloud-sim-uj/PyMPDATA/commit/1ad49aad3dcac7ec299637005f143ad363d6c7b6
Thank you!
