+++ paper/paper.md
@@ -77,7 +77,7 @@ Usage of ``PyMPDATA`` from ``Julia`` (https://julialang.org) and ``Matlab`` (htt
   is depicted in the PyMPDATA README file.
 The appendices of the present paper include ``Python``, ``Julia`` and ``Matlab`` minimal 
   code snippets covering steps needed to complete a basic ``PyMPDATA`` simulation
-  depicted in Fig. (\autoref{fig}) and based on Fig. 5 from @Arabas_et_al_2014.
+  depicted in \autoref{fig} and based on Figure 5 from @Arabas_et_al_2014.
 
 ![Visualisation of the initial condition (left) and simulation state after 75 timesteps (right) from the 2D simulation for which sample codes are given in the appendices, based on Fig. 5 from @Arabas_et_al_2014.\label{fig}](fig-crop.pdf)
 
@@ -109,8 +109,7 @@ The ``PyMPDATA``, the ``PyMPDATA-examples`` and the ``numba-mpi`` packages are a
 Development of all three packages is hosted on ``GitHub`` at: https://github.com/atmos-cloud-sim-uj/
   and continuous integration runs on ``Linux``, ``macOS`` and ``Windows`` using
   ``GitHub Actions`` and ``Appveyor`` platforms (the latter used for 32-bit runs on Windows).
-Auto-generated documentation sites built with ``pdoc3``  
-are hosted at
+Auto-generated documentation sites built with ``pdoc3`` are hosted at
   https://atmos-cloud-sim-uj.github.io/PyMPDATA/, 
   https://atmos-cloud-sim-uj.github.io/PyMPDATA-examples/, and 
   https://atmos-cloud-sim-uj.github.io/numba-mpi/.
@@ -182,7 +181,7 @@ Simulations are carried out on one, two or three threads on a machine with four 
 
 ![Comparison of wall-time measurements results for a 3D simulation using PyMPDATA with JIT disabled (red line) and enabled (connected points) corroborated against timings of analogous simulation performed with libmpdata++. Panel (a) presents scaling with the number of threads used, for the case of 16 by 16 by 16 domain. Panel (b) depicts scaling with domain size for simulations using three threads.\label{fig:perf}](fig-perf.pdf)
 
-Figure \autoref{fig:perf} (a) depicts wall-times measured with a domain of 16 by 16 by 16, and
+\autoref{fig:perf} (a) depicts wall-times measured with a domain of 16 by 16 by 16, and
   for: ``PyMPDATA`` with Numba JIT disabled (red line),
   ``libmpdata++`` (green connected points), and ``PyMPDATA`` with JIT enabled for both dynamic grid
   (i.e., grid extents specified at run-time, plotted with orange connected points) and
@@ -193,7 +192,7 @@ Comparison of ``PyMPDATA`` and ``libmpdata++`` reveals comparable performance an
   of threads with consistently shorter wall-times for ``PyMPDATA``, and a slight further improvement
   when switching from dynamic to static grid.
 
-Figure \autoref{fig:perf} (b) depicts wall-time dependence on the domain size for the case of three threads,
+\autoref{fig:perf} (b) depicts wall-time dependence on the domain size for the case of three threads,
   and confirms that the observed higher performance of ``PyMPDATA`` as compared with libmpdata++ can be observed 
   over a range of domain sizes starting from 16 by 16 by 16 up to 128 by 128 by 128.
 While more comprehensive tests and analyses would be needed to identify the cause of this superior
