### 2022-09-05 #3896 @arfon
@slayoo – can you merge this PR please? It fixes a few minor formatting issues: https://github.com/atmos-cloud-sim-uj/PyMPDATA/pull/337
