### 2022-09-05 #3896 @slayoo
@arfon Thank you!
Unfortunately, we have an issue with a non-breaking space in the title that is wrongly displayed on the JOSS website here:
![image](https://user-images.githubusercontent.com/89685/188433194-a3b895ed-f181-4f4f-a2cd-8c37406a74a8.png)
