### 2022-09-05 #3896 @arfon
Can you remove the &nbsp; from the title text please and I'll reprocess.
