### 2022-09-05 #3896 @arfon
@slayoo fixed in https://github.com/atmos-cloud-sim-uj/PyMPDATA/pull/338
