### 2022-09-05 #3896 @slayoo
BTW, we have an analogous problem in https://github.com/openjournals/joss-reviews/issues/4326
