### 2022-09-05 #3896 @arfon
OK, this should be fixed now.
