### 2022-09-06 #3896 @slayoo
@arfon I've noticed that there were three more nbsps in the author list that also go through into the paper metadata and JOSS website:

![image](https://user-images.githubusercontent.com/89685/188574978-f60daccc-31cd-4a70-89fa-81ed74eb025b.png)

Just removed it in the repo: https://github.com/atmos-cloud-sim-uj/PyMPDATA/commit/a35a159232f2021f31eab07bdc5178c534c2c848

Let me ask for one more regeneration of the metadata.
Thank you!
Sylwester
