### 2022-09-09 #3896 @slayoo
@arfon, please let us know if it would be possible to fix the remaining nbsp problems in the author list (already fixed in the project  repo) - see above, thanks, Sylwester
