### 2022-09-15 #3896 @xuanxu
@kthyng if the paper is already updated in the original repo, the `reaccept` command should regenerate the metadata
