### 2022-09-15 #3896 @slayoo
@kthyng it's now fixed:
![image](https://user-images.githubusercontent.com/89685/190449370-a5fa4df0-b774-47ba-9170-2bbe5040c2df.png)

Thank you for your help!
